pyproject.toml
[tool.poetry]
name = "teste-manoel"
version = "0.1.0"
description = ""
authors = ["Manoel Messias <manoelmsnsi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
sqlalchemy = "^2.0.31"
pydantic = {extras = ["email"], version = "^2.10.5"}
alembic = "^1.13.1"
python-dotenv = {extras = ["cli"], version = "^1.0.1"}
fastapi = "^0.111.0"
fastapi-pagination = "^0.12.25"
uvicorn = "^0.30.1"
asyncpg = "^0.30.0"
tenacity = "^9.0.0"
python-multipart = "^0.0.20"
pytz = "^2024.2"
pycryptodome = "^3.21.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
python-jose = "^3.3.0"
elastic-apm = "^6.23.0"
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
